Format-specific output for generating daily hashes of metric values (#1)
For each "application" ('a' field in the log) and for each day, there is a redis hash. This redis hash contains an entry for each metric ('m' field in the log), which is handled as a counter or as a discrete value (depending if 'v' or 'd' is used for the value).